Fix incremental compilation
During subsequent stage of incremental compilation not all files
may be available for indexing. Before overwriting the index file
this read its content and merge it with what was indexed in this stage.

There is one problem left with the classes which where removed.
Those will still be present and might trigger exception at run-time.

Fixes: #1